% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits_to_age_groups.R
\name{age_groups_to_limits}
\alias{age_groups_to_limits}
\title{Convert age groups to lower age limits}
\usage{
age_groups_to_limits(x)
}
\arguments{
\item{x}{age groups (a factor, as produced by \code{\link[=limits_to_age_groups]{limits_to_age_groups()}} or
\code{\link[=assign_age_groups]{assign_age_groups()}})}
}
\value{
a numeric vector of lower age limits
}
\description{
Inverse of \code{\link[=limits_to_age_groups]{limits_to_age_groups()}}. Extracts lower age limits from age group
labels.
}
\examples{
age_groups_to_limits(
  limits_to_age_groups(c(0, 5, 10), notation = "brackets")
)
}
