% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits_to_age_groups.R
\name{agegroups_to_limits}
\alias{agegroups_to_limits}
\title{Convert age groups to lower age limits (defunct)}
\usage{
agegroups_to_limits(x)
}
\arguments{
\item{x}{age groups (a factor, as produced by \code{\link[=limits_to_age_groups]{limits_to_age_groups()}} or
\code{\link[=assign_age_groups]{assign_age_groups()}})}
}
\value{
Always errors.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

\code{agegroups_to_limits()} was renamed to \code{\link[=age_groups_to_limits]{age_groups_to_limits()}} for naming
consistency.
}
\keyword{internal}
