% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check.r
\name{check.survey}
\alias{check.survey}
\title{Check that a survey fulfils all the requirements to work with the 'contact_matrix' function}
\usage{
\method{check}{survey}(survey, columns = FALSE, quiet = FALSE,
  error = FALSE, id.column = "part_id",
  participant.age.column = "part_age", country.column = "country",
  year.column = "year", contact.age.column = "cnt_age", ...)
}
\arguments{
\item{survey}{A list of 'participants' and 'contacts' (both data frames)}

\item{columns}{if given, a named character vector containing the name of the "id", "participant.age" and "contact.age" columns}

\item{quiet}{if TRUE, will not exit quietly if the test is passed.}

\item{error}{if TRUE, will stop if an error is found in the structure of the \code{participants} and \code{contacts} data frame}

\item{id.column}{the column in both the \code{participants} and \code{contacts} data frames that links contacts to participants}

\item{participant.age.column}{the column in the \code{participants} data frame containing participants' age}

\item{country.column}{the column in the \code{participants} data frame containing the country in which the participant was queried}

\item{year.column}{the column in the \code{participants} data frame containing the year in which the participant was queried}

\item{contact.age.column}{the column in the \code{contacts} data frame containing contacts' age}

\item{...}{ignored}

\item{warn}{don't stop on errors, only warn; useful to see all problems at once}
}
\value{
invisibly returns a character vector of the relevant columns
}
\description{
Check that a survey fulfils all the requirements to work with the 'contact_matrix' function
}
