% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sample-contact-ages.R
\name{check_min_n}
\alias{check_min_n}
\title{Validate the \code{min_n} argument of \code{\link[=sample_contact_ages]{sample_contact_ages()}}}
\usage{
check_min_n(min_n)
}
\arguments{
\item{min_n}{object to validate}
}
\value{
invisibly \code{NULL}; errors if \code{min_n} is not a positive integer
}
\description{
Validate the \code{min_n} argument of \code{\link[=sample_contact_ages]{sample_contact_ages()}}
}
\keyword{internal}
