% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_survey.r
\name{check_survey}
\alias{check_survey}
\title{Check that a survey fulfils all the requirements to work with the 'contact_matrix' function}
\usage{
check_survey(survey, quiet = FALSE)
}
\arguments{
\item{survey}{A list of 'participants' and 'contacts' (both data frames)}
}
\value{
TRUE if the test is passed, an error is thrown otherwise
}
\description{
Check that a survey fulfils all the requirements to work with the 'contact_matrix' function
}
