% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internet.r
\name{check_zenodo}
\alias{check_zenodo}
\title{Check the status of the Zenodo repository}
\usage{
check_zenodo()
}
\value{
HTTP status code, or 0 if internet down.
}
\description{
Check the status of the Zenodo repository
}
\author{
Sebastian Funk
}
\keyword{internal}
