% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite.r
\name{cite}
\alias{cite}
\alias{cite.survey}
\title{Citation for a survey}
\usage{
\method{cite}{survey}(x, quiet = FALSE, ...)
}
\arguments{
\item{x}{a character vector of surveys to cite}

\item{quiet}{if set to TRUE, do not print entry, just return bibentry object}

\item{...}{ignored}
}
\value{
citation as bibentry
}
\description{
Gets a full citation for a \code{\link{survey}}. If quiet is FALSE (default)
}
