% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite.r
\name{cite.survey}
\alias{cite.survey}
\title{Print citation for a survey}
\usage{
\method{cite}{survey}(survey, quiet = FALSE, ...)
}
\arguments{
\item{survey}{a character vector of surveys to cite}

\item{quiet}{if set to TRUE, do not print entry, just return bibentry object}
}
\value{
citation as bibentry
}
\description{
Print citation for a survey
}
