% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\alias{clean.contact_survey}
\title{Clean contact survey data}
\usage{
\method{clean}{contact_survey}(
  x,
  participant_age_column = "part_age",
  ...,
  participant.age.column = deprecated()
)
}
\arguments{
\item{x}{A \code{\link[=survey]{survey()}} object}

\item{participant_age_column}{the column in \code{x$participants}
containing participants' age}

\item{...}{ignored}

\item{participant.age.column}{\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}
Use \code{participant_age_column} instead.}
}
\value{
a cleaned survey in the correct format
}
\description{
Cleans survey data to work with the 'contact_matrix' function
}
\examples{
data(polymod)
cleaned <- clean(polymod) # not really necessary, polymod is clean
}
