% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_matrix.r
\name{contact_matrix_multi}
\alias{contact_matrix_multi}
\title{Generate multiple contact matrices using bootstrapping}
\usage{
contact_matrix_multi(n, ...)
}
\arguments{
\item{n}{number contact matrices to generate}

\item{...}{parameters to pass to \link{contact_matrix}}
}
\value{
a list of contact matrices, and the underlying demography of the surveyed population
}
\description{
This function can be used to take into account uncertainty in generating
the contact matrices stemming from the finite underlying population. It
does so by calling the \link{contact_matrix} function multiple times. If any
of the arguments passed to \link{contact_matrix} refer to sampling (especially
'sample.participants') then the resulting contact matrices will differ and
can be used to generate bootstrap confidence intervals, or multiple
verisons of contact matrices consistsent with the data.
}
\author{
Sebastian Funk
}
