% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-contact-survey.R
\name{copy.contact_survey}
\alias{copy.contact_survey}
\title{Deep copy a contact survey}
\usage{
\method{copy}{contact_survey}(x)
}
\arguments{
\item{x}{a \code{contact_survey} object}
}
\value{
a deep copy of the survey
}
\description{
Creates a deep copy of a \code{contact_survey} object, including its
\code{participants} and \code{contacts} data.tables.
}
