% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subset-contact-survey.R
\name{copy_survey}
\alias{copy_survey}
\title{Deep copy a contact survey}
\usage{
copy_survey(survey)
}
\arguments{
\item{survey}{a \code{contact_survey} object}
}
\value{
a deep copy of the survey
}
\description{
Creates a deep copy of a \code{contact_survey} object, including its
\code{participants} and \code{contacts} data.tables.
}
\keyword{internal}
