% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey.r
\name{.get_survey}
\alias{.get_survey}
\title{Internal function to get survey data}
\usage{
.get_survey(survey, ...)
}
\arguments{
\item{survey}{a DOI or url to get the survey from, or a \code{\link[=survey]{survey()}} object (in which case only cleaning is done).}

\item{...}{options for \code{\link[=clean]{clean()}}, which is called at the end of this}
}
\description{
Internal function to get survey data
}
\keyword{internal}
