% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{download_survey}
\alias{download_survey}
\title{Download a survey from its Zenodo repository}
\usage{
download_survey(survey, dir = NULL, sleep = 1)
}
\arguments{
\item{survey}{a URL (see \code{contactsurveys::list_surveys()})}

\item{dir}{a directory to save the files to; if not given, will save to a
temporary directory}

\item{sleep}{time to sleep between requests to avoid overloading the server
(passed on to \code{\link[base]{Sys.sleep}})}
}
\value{
Always errors.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

\code{download_survey()} is defunct. Use \code{contactsurveys::download_survey()}
instead.
}
\examples{
\dontrun{
peru_survey <- contactsurveys::download_survey(
  "https://doi.org/10.5281/zenodo.1095664"
)
}
}
\seealso{
load_survey
}
