% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internet.r
\name{ensure_zenodo_available}
\alias{ensure_zenodo_available}
\title{Helper function to ensure Zenodo can be accessed}
\usage{
ensure_zenodo_available()
}
\description{
Helper function to ensure Zenodo can be accessed
}
\author{
Sebastian Funk
}
\keyword{internal}
