% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_survey.R
\name{filter_survey}
\alias{filter_survey}
\title{Filter participant and contact data}
\usage{
filter_survey(x, ...)
}
\arguments{
\item{x}{a \verb{<contact_survey>} object}

\item{...}{any filters to apply to the data, given as list of the
form (column=filter_value) - only contacts that have 'filter_value' in
'column' will be considered. If multiple filters are given, they are all
applied independently and in the sequence given.}
}
\value{
a filtered \verb{<contact_survey>} object
}
\description{
Filter participant and contact data
}
\examples{
## filter the survey data for home contacts of participants from Germany
filter_survey(polymod, country = "Germany", cnt_home = 1)
}
