% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-survey-utils.R
\name{get_mergeable_files}
\alias{get_mergeable_files}
\title{Identify which survey files share columns with a main table}
\usage{
get_mergeable_files(survey_files, contact_data, main_cols)
}
\description{
Identify which survey files share columns with a main table
}
\keyword{internal}
