% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_survey}
\alias{get_survey}
\title{Get a survey, either from its Zenodo repository, a set of
files, or a survey variable}
\usage{
get_survey(survey, clear_cache = FALSE, ...)
}
\arguments{
\item{survey}{a DOI or url to get the survey from, or a survey object}

\item{clear_cache}{logical, whether to clear the cache before downloading
the survey}

\item{...}{currently unused}
}
\value{
Always errors.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

\code{get_survey()} is defunct. Use \code{contactsurveys::download_survey()} and then
\code{\link[=load_survey]{load_survey()}} instead.
}
\examples{
\dontrun{
peru_doi <- "https://doi.org/10.5281/zenodo.1095664"
peru_survey <- contactsurveys::download_survey(peru_doi)
peru_data <- load_survey(peru_survey)
}
}
