% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact-matrix-result.R
\name{is_contact_matrix}
\alias{is_contact_matrix}
\title{Test whether an object is a contact_matrix}
\usage{
is_contact_matrix(x)
}
\arguments{
\item{x}{object to test}
}
\value{
logical
}
\description{
Test whether an object is a contact_matrix
}
\keyword{internal}
