% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{is_doi}
\alias{is_doi}
\title{Checks if a character string is a DOI}
\usage{
is_doi(x)
}
\arguments{
\item{x}{Character vector; the string or strings to check}
}
\value{
Logical; \code{TRUE} if \code{x} is a DOI, \code{FALSE} otherwise
}
\description{
Checks if a character string is a DOI
}
\author{
Sebastian Funk
}
