% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits_to_age_groups.R
\name{limits_to_age_groups}
\alias{limits_to_age_groups}
\title{Convert lower age limits to age groups.}
\usage{
limits_to_age_groups(
  x,
  limits = sort(unique(x)),
  notation = c("dashes", "brackets")
)
}
\arguments{
\item{x}{age limits to transform}

\item{limits}{lower age limits; if not given, will use all limits in \code{x}}

\item{notation}{whether to use bracket notation, e.g. [0,4) or dash
notation, e.g. 0-4)}
}
\value{
Age groups as specified in \code{notation}
}
\description{
Mostly used for plot labelling
}
\examples{
limits_to_age_groups(c(0, 5, 10))
}
