% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_surveys.r
\name{merge_surveys}
\alias{merge_surveys}
\title{Merge multiple surveys into one}
\usage{
merge_surveys(surveys)
}
\arguments{
\item{surveys}{surveys to merge, given as a list of surveys compatible with the \code{survey} parameter in \code{\link{contat_mtarix}}.}
}
\value{
a survey in the correct format
}
\description{
Merge multiple surveys into one
}
