% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_survey.R
\name{new_survey}
\alias{new_survey}
\title{Contact survey}
\usage{
new_survey(participants, contacts, reference = NULL)
}
\arguments{
\item{participants}{a \code{data.frame} containing information on participants}

\item{contacts}{a \code{data.frame} containing information on contacts}

\item{reference}{a \code{list} containing information needed to reference the survey, in particular it can contain$a "title", "bibtype", "author", "doi", "publisher", "note", "year"}
}
\value{
a new survey object
}
\description{
Deprecated. A \code{survey} object contains the results of a contact survey. In particular, it contains two data frames called \code{participants} and \code{contacts} that are linked by a column specified as \code{id.column}
}
\examples{
data(polymod)
survey <- new_survey(polymod$participants, polymod$contacts)
}
\author{
Sebastian Funk
}
