% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{normalise_country_names}
\alias{normalise_country_names}
\title{Normalise country names}
\usage{
normalise_country_names(countries)
}
\arguments{
\item{countries}{A vector of country names or codes}
}
\value{
A character vector of normalised country names
}
\description{
Uses the countrycode package to standardise country names. This handles
2-letter ISO codes, 3-letter ISO codes, and full country names, converting
them all to standardised country names.
}
\keyword{internal}
