% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasets.r
\docType{data}
\name{polymod}
\alias{polymod}
\title{Social contact data from 8 European countries}
\format{A list of two data frames:
\describe{
  \item{participants}{the study participant, with age, country, year and day
  of the week (starting with 1 = Monday)}
  \item{contacts}{weight of the diamond, in carats}
  ...
}}
\source{
\url{http://dx.doi.org/10.1371/journal.pmed.0050074}
}
\usage{
polymod
}
\description{
A dataset containing social mixing diary data from 8 European countries:
Belgium, Germany, Finland, Great Britain, Italy, Luxembourg,
The Netherlands and Poland.
The Data are fully described in Mossong J, Hens N, Jit M, Beutels P, Auranen
K, Mikolajczyk R, et al. (2008) Social Contacts and Mixing Patterns Relevant
to the Spread of Infectious Diseases. PLoS Med 5(3): e74.
}
\keyword{datasets}
