% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_age_groups.R
\name{reduce_agegroups}
\alias{reduce_agegroups}
\title{Reduce the number of age groups (defunct)}
\usage{
reduce_agegroups(x, limits)
}
\arguments{
\item{x}{vector of limits}

\item{limits}{new limits}
}
\value{
Always errors.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

\code{reduce_agegroups()} was renamed to \code{\link[=reduce_age_groups]{reduce_age_groups()}} for naming
consistency.
}
\keyword{internal}
