% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey.R
\name{survey}
\alias{survey}
\title{Contact survey (defunct)}
\usage{
survey(participants, contacts, reference = NULL)
}
\arguments{
\item{participants}{a \code{data.frame} containing information on participants}

\item{contacts}{a \code{data.frame} containing information on contacts}

\item{reference}{a \code{list} containing information needed to
reference the survey, in particular it can contain a "title",
"bibtype", "author", "doi", "publisher", "note", "year"}
}
\value{
Always errors.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

\code{survey()} is defunct. Use \code{\link[=as_contact_survey]{as_contact_survey()}} instead.
}
\seealso{
\code{\link[=as_contact_survey]{as_contact_survey()}}
}
\author{
Sebastian Funk
}
