% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey_citation.r
\name{survey_citation}
\alias{survey_citation}
\title{Print citation for a survey}
\usage{
survey_citation(survey, quiet = FALSE)
}
\arguments{
\item{survey}{a character vector of surveys to cite}

\item{quiet}{if set to TRUE, do not print entry, just return bibentry object}
}
\value{
citation as bibentry
}
\description{
Print citation for a survey
}
