% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{survey_countries}
\alias{survey_countries}
\title{List all countries contained in a survey}
\usage{
survey_countries(survey, country.column = "country", ...)
}
\arguments{
\item{survey}{a DOI or url to get the survey from, or a survey object}

\item{country.column}{column in the survey indicating the country}

\item{...}{currently unused}
}
\value{
Always errors.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}
}
\details{
\code{survey_countries()} is defunct. Use \code{contactsurveys::download_survey()}
and \code{\link[=load_survey]{load_survey()}} and then explore the country column yourself.
}
\examples{
\dontrun{
doi_peru <- "10.5281/zenodo.1095664" # nolint
# download the data with the contactsurveys package
peru_survey <- contactsurveys::download_survey(doi_peru)
# load the survey with socialmixr
peru_data <- socialmixr::load_survey(peru_survey)
# find the unique country - assuming your data has a "country" column:
unique(peru_data$participants$country)
}
}
