% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-process-age.R
\name{survey_process_ages}
\alias{survey_process_ages}
\title{Process ages in survey data}
\usage{
survey_process_ages(
  survey,
  age_limits = NULL,
  missing_participant_age = c("remove", "keep"),
  missing_contact_age = c("remove", "sample", "keep", "ignore")
)
}
\arguments{
\item{survey}{a \code{\link[=survey]{survey()}} object}

\item{age_limits}{lower limits of the age groups over which to construct
the matrix. Defaults to NULL. If NULL, it will construct age limits based
upon age ranges in participants data of the survey.}

\item{missing_participant_age}{if set to "remove" (default), participants
without age information are removed; if set to "keep", participants with
missing age are kept and treated as a separate age group}

\item{missing_contact_age}{if set to "remove" (default), participants that
have contacts without age information are removed; if set to "sample",
contacts without age information are sampled from all the contacts of
participants of the same age group; if set to "keep", contacts with missing
age are kept and treated as a separate age group; if set to "ignore",
contact with missing age are ignored in the contact analysis}
}
\value{
The survey object with processed age data.
}
\description{
This function deals with age ranges and missing data in survey data. It
adjusts the age group breaks to the lower and upper ages in the survey,
and processes contact age ranges and drops missing values. We suggest you
impute missing values prior to using \code{survey_process_ages}, using
\code{\link[=survey_impute_ages]{survey_impute_ages()}}.
}
\examples{
polymod_imputed_processed <- polymod |> survey_impute_ages() |> survey_process_ages()
polymod_imputed_processed
polymod_processed <- polymod |> survey_process_ages()
polymod_processed
}
