% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load-survey-utils.R
\name{try_merge_one_file}
\alias{try_merge_one_file}
\title{Try merging a single additional file into a main survey table}
\usage{
try_merge_one_file(
  file,
  type,
  main_survey,
  contact_data,
  participant_key = NULL,
  call = rlang::caller_env()
)
}
\value{
A list with components: merged (data.table or NULL), detected_key
(character vector or NULL), file (the file that was merged, or NULL).
}
\description{
Try merging a single additional file into a main survey table
}
\keyword{internal}
