% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact-age-distribution.R
\name{validate_age_distribution}
\alias{validate_age_distribution}
\title{Validate an age distribution data.frame}
\usage{
validate_age_distribution(x)
}
\arguments{
\item{x}{object to validate}
}
\value{
\code{x} invisibly, or errors
}
\description{
Validate an age distribution data.frame
}
\keyword{internal}
