% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internet.r
\name{zenodo_available}
\alias{zenodo_available}
\title{Check if the Zenodo repository is available}
\usage{
zenodo_available()
}
\value{
TRUE or FALSE, depending on whether Zenodo can be accessed.
}
\description{
Check if the Zenodo repository is available
}
\author{
Sebastian Funk
}
\keyword{internal}
