% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits_to_agegroups.R
\name{agegroups_to_limits}
\alias{agegroups_to_limits}
\title{Convert age groups to lower age limits}
\usage{
agegroups_to_limits(x)
}
\arguments{
\item{x}{age groups (a factor, as produced by \code{\link[=limits_to_agegroups]{limits_to_agegroups()}} or
\code{\link[=assign_age_groups]{assign_age_groups()}})}
}
\value{
a numeric vector of lower age limits
}
\description{
Inverse of \code{\link[=limits_to_agegroups]{limits_to_agegroups()}}. Extracts lower age limits from age group
labels.
}
\examples{
agegroups_to_limits(limits_to_agegroups(c(0, 5, 10), notation = "brackets"))
}
