% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_contact_survey.R
\name{as_contact_survey}
\alias{as_contact_survey}
\title{Check contact survey data}
\usage{
as_contact_survey(
  x,
  id.column = "part_id",
  country.column = "country",
  year.column = "year"
)
}
\arguments{
\item{x}{list containing
\itemize{
\item an element named 'participants', a data frame containing participant
information
\item an element named 'contacts', a data frame containing contact information
}}

\item{id.column}{the column in both the \code{participants} and \code{contacts} data frames that links contacts to participants}

\item{country.column}{the column in the \code{participants} data frame containing the country in which the participant was queried}

\item{year.column}{the column in the \code{participants} data frame containing the year in which the participant was queried
\itemize{
\item (optionally) an element named 'reference, a \code{\link[utils:bibentry]{utils::bibentry()}} object
}}
}
\value{
invisibly returns a character vector of the relevant columns
}
\description{
Checks that a survey fulfills all the requirements to work with the 'contact_matrix' function
}
\examples{
data(polymod)
check(polymod)
}
