% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/as_contact_survey.R
\name{as_contact_survey}
\alias{as_contact_survey}
\title{Check contact survey data}
\usage{
as_contact_survey(
  x,
  reference = NULL,
  id.column = "part_id",
  country.column = "country",
  year.column = "year"
)
}
\arguments{
\item{x}{list containing
\itemize{
\item an element named 'participants', a data frame containing participant
information
\item an element named 'contacts', a data frame containing contact information
\item (optionally) an element named 'reference, a list containing information
information needed to reference the survey, in particular it can contain$a
"title", "bibtype", "author", "doi", "publisher", "note", "year"
}}

\item{reference}{a \code{list} containing information needed to reference the survey, in particular it can contain$a "title", "bibtype", "author", "doi", "publisher", "note", "year"}

\item{id.column}{the column in both the \code{participants} and \code{contacts} data frames that links contacts to participants}

\item{country.column}{the column in the \code{participants} data frame containing the country in which the participant was queried}

\item{year.column}{the column in the \code{participants} data frame containing the year in which the participant was queried}

\item{participant.age.column}{the column in the \code{participants} data frame containing participants' age; if this does not exist, at least columns "..._exact", "..._est_min" and "..._est_max" must (see the \code{estimated.participant.age} option in \code{\link[=contact_matrix]{contact_matrix()}})}

\item{contact.age.column}{the column in the \code{contacts} data frame containing the age of contacts; if this does not exist, at least columns "..._exact", "..._est_min" and "..._est_max" must (see the \code{estimated.contact.age} option in \code{\link[=contact_matrix]{contact_matrix()}})}

\item{...}{ignored}
}
\value{
invisibly returns a character vector of the relevant columns
}
\description{
Checks that a survey fulfills all the requirements to work with the 'contact_matrix' function
}
\examples{
data(polymod)
check(polymod)
}
