% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite.r
\name{cite}
\alias{cite}
\alias{cite.survey}
\title{Citation for a survey}
\usage{
\method{cite}{survey}(x, ...)
}
\arguments{
\item{x}{a character vector of surveys to cite}

\item{...}{ignored}
}
\value{
citation as bibentry
}
\description{
Gets a full citation for a \code{\link[=survey]{survey()}}. If quiet is FALSE (default)
}
\examples{
data(polymod)
cite(polymod)
}
