% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clean.R
\name{clean}
\alias{clean}
\alias{clean.contact_survey}
\title{Clean contact survey data}
\usage{
\method{clean}{contact_survey}(x, country.column = "country", participant.age.column = "part_age", ...)
}
\arguments{
\item{x}{A \code{\link[=survey]{survey()}} object}

\item{country.column}{the name of the country in which the survey participant was interviewed}

\item{participant.age.column}{the column in \code{x$participants} containing participants' age}

\item{...}{ignored}
}
\value{
a cleaned survey in the correct format
}
\description{
Cleans survey data to work with the 'contact_matrix' function
}
\examples{
data(polymod)
cleaned <- clean(polymod) # not really necessary as the 'polymod' data set has already been cleaned
}
