% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_matrix.r
\name{contact_matrix}
\alias{contact_matrix}
\title{Generate a contact matrix from diary survey data}
\usage{
contact_matrix(survey = "polymod", countries = c(), survey.pop, age.limits,
  filter, n = 1, bootstrap, counts = FALSE, symmetric = FALSE,
  split = FALSE, weigh.dayofweek = FALSE, weights = c(), quiet = FALSE)
}
\arguments{
\item{survey}{either a (case-insensitive) survey name ("POLYMOD") or a list of 'participants' and 'contacts' (both data frames) to sample from}

\item{countries}{limit to one or more countries; if not given, will use all countries in the survey}

\item{survey.pop}{survey population -- either a data frame with columns 'lower.age.limit' and 'population', or a character vector giving the name(s) of a country or countries from the list that can be obtained via \code{wpp_countries}; if not given, will use the country populations from the chosen countries, or all countries in the survey if \code{countries} is not given}

\item{age.limits}{lower age limits of age groups to extract}

\item{filter}{any filters to apply to the data, given as list of the form (column=filter_value) - only contacts that have 'filter_value' in 'column' will be considered}

\item{n}{number of matrices to sample}

\item{bootstrap}{whether to sample using a bootstrap; by default, will use bootstrap if n > 1}

\item{counts}{whether to return counts (instead of means)}

\item{symmetric}{whether to make matrix symmetric}

\item{split}{whether to split the number of contacts and assortativity}

\item{weigh.dayofweek}{whether to weigh the day of the week (weight 5 for weekdays ans 2 for weekends)}

\item{weights}{columns that contain weights}

\item{quiet}{if TRUE, suppress messages}
}
\value{
a list of sampled contact matrices, and the underlying demography of the surveyed population
}
\description{
Samples a contact survey using a bootstrap
}
\examples{
m <- contact_matrix()
m <- contact_matrix(n = 5)
m <- contact_matrix(split = TRUE)
m <- contact_matrix(survey = "POLYMOD", countries = "United Kingdom", age.limits = c(0, 1, 5, 15))
}
\author{
Sebastian Funk
}
