% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_matrix.r
\name{contact_matrix}
\alias{contact_matrix}
\title{Samples polymod using weights following Baguelin et al. (2013), but
using a bootstrap; first, contacts and ages are sampled}
\usage{
contact_matrix(n = 1, age.limits, survey = "POLYMOD", countries, survey.pop,
  bootstrap = FALSE, symmetric = TRUE, normalise = FALSE, split = FALSE,
  add.weights = c(), part.age.column = "participant_age",
  contact.age.column = "cnt_age_mean", id.column = "global_id",
  dayofweek.column = "day_of_week", country.column = "country",
  year.column = "year")
}
\arguments{
\item{n}{number of matrices to sample}

\item{age.limits}{Lower limits of the age groups; if not given, will use 5 year age limits as in the population data}

\item{survey}{either a (case-insensitive) survey name ("POLYMOD") or a list of 'participants' and 'contacts' (both data frames) to sample from}

\item{countries}{limit to one or more countries; if not given, will use all countries in the survey}

\item{survey.pop}{survey population -- either a data frame with columns lower.age.limit and population, or a character vector giving the name(s) to use with the 2013 WHO population; if not given, will use the country populations from the desired countries, or all countries in the survey if \code{countries} is not given}

\item{bootstrap}{whether to sample using a bootstrap; will be set to TRUE if n > 1}

\item{symmetric}{whether to make matrix symmetric}

\item{normalise}{whether to normalise to eigenvalue 1}

\item{split}{whether to split the number of contacts and assortativity}

\item{add.weights}{additional weight columns (e.g., minutes etc}

\item{part.age.column}{column indicating age in participant data}

\item{contact.age.column}{column indicating age in contact data}

\item{id.column}{column to match participants with contacts}

\item{dayofweek.column}{column indicating the day of the week}

\item{country.column}{column indicating the country}

\item{year.column}{column indicating the year}
}
\value{
a list of sampled contact matrices, and the underlying demography
}
\description{
Samples polymod using weights following Baguelin et al. (2013), but
using a bootstrap; first, contacts and ages are sampled
}
\author{
Sebastian Funk
}

