% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact_matrix.r
\name{contact_matrix}
\alias{contact_matrix}
\title{Generate a contact matrix from diary survey data}
\usage{
contact_matrix(survey = "POLYMOD", countries, survey.pop, age.limits, filter,
  n = 1, bootstrap = FALSE, symmetric = TRUE, normalise = FALSE,
  split = FALSE, weights = c(), part.age.column = "participant_age",
  contact.age.column = "cnt_age_mean", id.column = "global_id",
  dayofweek.column = "day_of_week", country.column = "country",
  year.column = "year", quiet = FALSE, ...)
}
\arguments{
\item{survey}{either a (case-insensitive) survey name ("POLYMOD") or a list of 'participants' and 'contacts' (both data frames) to sample from}

\item{countries}{limit to one or more countries; if not given, will use all countries in the survey}

\item{survey.pop}{survey population -- either a data frame with columns 'lower.age.limit' and 'population', or a character vector giving the name(s) of a country or countries from the list that can be obtained via \code{wpp_countries}; if not given, will use the country populations from the chosen countries, or all countries in the survey if \code{countries} is not given}

\item{age.limits}{lower age limits of age groups to extract}

\item{filter}{any filters to apply to the data, given as list of the form (column=filter_value) - only contacts that have 'filter_value' in 'column' will be considered}

\item{n}{number of matrices to sample}

\item{bootstrap}{whether to sample using a bootstrap; will be set to TRUE if n > 1}

\item{symmetric}{whether to make matrix symmetric}

\item{normalise}{whether to normalise to eigenvalue 1}

\item{split}{whether to split the number of contacts and assortativity}

\item{weights}{columns that contain weights}

\item{part.age.column}{column indicating age in participant data}

\item{contact.age.column}{column indicating age in contact data}

\item{id.column}{column to match participants with contacts}

\item{dayofweek.column}{column indicating the day of the week}

\item{country.column}{column indicating the country}

\item{year.column}{column indicating the year}

\item{quiet}{if TRUE, suppress messages}

\item{...}{further parameters for 'pop_age'}
}
\value{
a list of sampled contact matrices, and the underlying demography of the surveyed population
}
\description{
Samples a contact survey using a bootstrap
}
\examples{
m <- contact_matrix()
m <- contact_matrix(n = 5)
m <- contact_matrix(normalise = TRUE)
m <- contact_matrix(normalise = TRUE, split = TRUE)
m <- contact_matrix(survey = "POLYMOD", countries = "United Kingdom", age.limits = c(0, 1, 5, 15))
}
\author{
Sebastian Funk
}
