% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecate-args.R
\name{deprecate_arg}
\alias{deprecate_arg}
\title{Handle deprecated argument}
\usage{
deprecate_arg(old_arg, new_arg, old_name, new_name, fn_name, version = "0.5.0")
}
\arguments{
\item{old_arg}{the deprecated argument value}

\item{new_arg}{the new argument value}

\item{old_name}{the old argument name (with dot)}

\item{new_name}{the new argument name (with underscore)}

\item{fn_name}{the function name}

\item{version}{the version when deprecated}
}
\value{
\code{new_arg}, or \code{old_arg} if it is supplied, in which case a
deprecation warning is issued via \code{\link[lifecycle:deprecate_soft]{lifecycle::deprecate_warn()}}.
}
\description{
Handle deprecated argument
}
\keyword{internal}
