% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download_survey.r
\name{download_survey}
\alias{download_survey}
\title{Download a survey from its Zenodo repository}
\usage{
download_survey(survey, dir = NULL)
}
\arguments{
\item{survey}{a URL (see \code{\link[=list_surveys]{list_surveys()}})}

\item{dir}{a directory to save the files to; if not given, will save to a temporary directory}
}
\value{
a vector of filenames that can be used with \link{load_survey}
}
\description{
Downloads survey data
}
\examples{
\dontrun{
list_surveys()
peru_survey <- download_survey("https://doi.org/10.5281/zenodo.1095664")
}
}
