% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_citation}
\alias{get_citation}
\title{Citation for a survey}
\usage{
get_citation(x)
}
\arguments{
\item{x}{a character vector of surveys to cite}
}
\value{
citation as bibentry
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

\code{get_citation()} is defunct. Use \code{contactsurveys::get_citation()} instead.
}
\examples{
# we recommend using the contactsurveys package for get_citation()
\dontrun{
data(polymod)
citation <- contactsurveys::get_citation(polymod)
print(citation)
print(citation, style = "bibtex")
}
}
