% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cite.r
\name{get_citation}
\alias{get_citation}
\title{Citation for a survey}
\usage{
get_citation(x)
}
\arguments{
\item{x}{a character vector of surveys to cite}
}
\value{
citation as bibentry
}
\description{
Gets a full citation for a \code{\link[=survey]{survey()}}.
}
\examples{
data(polymod)
citation <- get_citation(polymod)
print(citation)
print(citation, style = "bibtex")
}
