% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey.r
\name{get_survey}
\alias{get_survey}
\title{Get a survey from a variable, either from internal data or passed as an argument}
\usage{
get_survey(survey)
}
\arguments{
\item{survey}{either a (case-insensitive) survey name ("POLYMOD") or a list of 'participants' and 'contacts' (both data frames) to sample from}
}
\description{
Get a survey from a variable, either from internal data or passed as an argument
}
