% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_survey.r
\name{get_survey}
\alias{get_survey}
\title{Get a survey, either from its Zenodo repository, a set of files, or a survey variable}
\usage{
get_survey(survey, ...)
}
\arguments{
\item{survey}{a DOI or url to get the survey from, or a \code{\link[=survey]{survey()}} object (in which case only cleaning is done).}

\item{...}{options for \code{\link[=clean]{clean()}}, which is called at the end of this}
}
\value{
a survey in the correct format
}
\description{
Downloads survey data, or extracts them from files, and returns a clean data set.
}
\examples{
\dontrun{
list_surveys()
peru_survey <- get_survey("https://doi.org/10.5281/zenodo.1095664")
}
}
