% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/survey-impute-ages.R
\name{impute_ages}
\alias{impute_ages}
\alias{survey_impute_ages}
\title{Impute missing ages in survey data}
\usage{
impute_ages(
  survey,
  missing_participant_age = c("mean", "sample", "missing"),
  missing_contact_age = c("mean", "sample", "missing")
)

survey_impute_ages(
  survey,
  missing_participant_age = c("mean", "sample", "missing"),
  missing_contact_age = c("mean", "sample", "missing")
)
}
\arguments{
\item{survey}{A survey object.}

\item{missing_participant_age}{if set to "mean" (default), people whose ages
are given as a range (in columns named "..._est_min" and "..._est_max") but
not exactly (in a column named "..._exact") will have their age set to the
mid-point of the range; if set to "sample", the age will be sampled from
the range; if set to "missing", age ranges will be treated as missing}

\item{missing_contact_age}{if set to "mean" (default), contacts whose ages
are given as a range (in columns named "..._est_min" and "..._est_max") but
not exactly (in a column named "..._exact") will have their age set to the
mid-point of the range; if set to "sample", the age will be sampled from
the range; if set to "missing", age ranges will be treated as missing}
}
\value{
The modified survey object with imputed ages for participants and contacts.
}
\description{
This function imputes the participant and contact data in a \code{\link[=survey]{survey()}}
object. You can impute these values yourself with
\code{\link[=impute_participant_ages]{impute_participant_ages()}} and \code{\link[=impute_contact_ages]{impute_contact_ages()}}. We recommend
imputing before processing ages with \code{\link[=process_ages]{process_ages()}}.
}
