% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/limits_to_age_groups.R
\name{limits_to_agegroups}
\alias{limits_to_agegroups}
\title{Convert lower age limits to age groups (deprecated)}
\usage{
limits_to_agegroups(
  x,
  limits = sort(unique(x)),
  notation = c("dashes", "brackets")
)
}
\arguments{
\item{x}{age limits to transform}

\item{limits}{lower age limits; if not given, will use all limits in \code{x}}

\item{notation}{whether to use bracket notation, e.g. [0,4) or dash
notation, e.g. 0-4)}
}
\value{
Age groups as specified in \code{notation}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{limits_to_agegroups()} was renamed to \code{\link[=limits_to_age_groups]{limits_to_age_groups()}} for naming
consistency.
}
\keyword{internal}
