% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{list_surveys}
\alias{list_surveys}
\title{List all surveys available for download}
\usage{
list_surveys(clear_cache = FALSE)
}
\arguments{
\item{clear_cache}{logical, whether to clear the cache before downloading
the survey; by default, the cache is not cleared and so multiple calls of
this function to access the same survey will not result in repeated
downloads.}
}
\value{
Always errors.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#defunct}{\figure{lifecycle-defunct.svg}{options: alt='[Defunct]'}}}{\strong{[Defunct]}}

\code{list_surveys()} is defunct. Use \code{contactsurveys::list_surveys()} instead.
}
\examples{
# we recommend using the contactsurveys package now for listing surveys.
\dontrun{
contactsurveys::list_surveys()
}
}
