% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.r
\name{list_surveys}
\alias{list_surveys}
\title{List all surveys available for download}
\usage{
list_surveys()
}
\value{
character vector of surveys
}
\description{
List all surveys available for download
}
\examples{
list_surveys()
}
