% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.R
\name{list_surveys}
\alias{list_surveys}
\title{List all surveys available for download}
\usage{
list_surveys(clear_cache = FALSE)
}
\arguments{
\item{clear_cache}{logical, whether to clear the cache before downloading the survey; by default, the cache is not cleared and so multiple calls of this function to access the same survey will not result in repeated downloads}
}
\value{
character vector of surveys
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{list_surveys()} has been deprecated in favour of
\code{\link[contactsurveys:list_surveys]{contactsurveys::list_surveys()}}.
}
\examples{
\dontrun{
list_surveys()
}
}
