% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_survey.r
\name{load_survey}
\alias{load_survey}
\title{Load a survey from local files}
\usage{
load_survey(files, ...)
}
\arguments{
\item{files}{a vector of file names as returned by \code{\link[=download_survey]{download_survey()}}}

\item{...}{options for \code{\link[=clean]{clean()}}, which is called at the end of this}
}
\value{
a survey in the correct format
}
\description{
Loads a survey from a local file system. Tables are expected as csv files, and a reference (if present) as JSON.
}
\examples{
\dontrun{
list_surveys()
peru_files <- download_survey("https://doi.org/10.5281/zenodo.1095664")
peru_survey <- load_survey(peru_files)
}
}
