% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pop_age.r
\name{pop_age}
\alias{pop_age}
\title{Get age-specific population data according to the World Population Prospects 2015 edition}
\usage{
pop_age()
}
\value{
data frame of age-specific population data
}
\description{
This uses data from the \code{wpp2015} package but combines male and female,
and converts age groups to lower age limits
}

