% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_age_groups.R
\name{reduce_agegroups}
\alias{reduce_agegroups}
\title{Reduce the number of age groups (deprecated)}
\usage{
reduce_agegroups(x, limits)
}
\arguments{
\item{x}{vector of limits}

\item{limits}{new limits}
}
\value{
vector with the new age groups
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{reduce_agegroups()} was renamed to \code{\link[=reduce_age_groups]{reduce_age_groups()}} for naming
consistency.
}
\keyword{internal}
