% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reduce_agegroups.r
\name{reduce_agegroups}
\alias{reduce_agegroups}
\title{Reduce the number of age groups given a broader set of limits}
\usage{
reduce_agegroups(x, limits)
}
\arguments{
\item{x}{vector of limits}

\item{limits}{new limits}
}
\value{
vector with the new agegroups
}
\description{
Operates on lower limits
}
