% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contact-matrix-utils.R
\name{sample_from_age_distribution}
\alias{sample_from_age_distribution}
\title{Sample ages from a distribution within \verb{[min, max]} bands}
\usage{
sample_from_age_distribution(mins, maxs, distribution)
}
\arguments{
\item{mins}{integer vector of lower bounds}

\item{maxs}{integer vector of upper bounds}

\item{distribution}{data.frame with \code{age} and \code{proportion} columns}
}
\value{
integer vector of sampled ages
}
\description{
Sample ages from a distribution within \verb{[min, max]} bands
}
\keyword{internal}
