% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/postprocess-matrix.R
\name{split_matrix}
\alias{split_matrix}
\title{Decompose a contact matrix into mean contacts, normalisation and
assortativity}
\usage{
split_matrix(x, survey_pop, ...)
}
\arguments{
\item{x}{a list as returned by \code{\link[=compute_matrix]{compute_matrix()}}, with elements \code{matrix}
and \code{participants}}

\item{survey_pop}{a data frame with columns \code{lower.age.limit} and
\code{population} (e.g. from \code{\link[=wpp_age]{wpp_age()}})}

\item{...}{passed to \code{\link[=pop_age]{pop_age()}} for interpolation}
}
\value{
\code{x} with \verb{$matrix} replaced by the assortativity matrix, plus
additional elements \verb{$mean.contacts}, \verb{$normalisation}, and \verb{$contacts}
}
\description{
Splits the contact matrix into the mean number of contacts across the whole
population (\code{mean.contacts}), a normalisation constant (\code{normalisation}),
age-specific contact rates (\code{contacts}), and an assortativity matrix
(replacing \verb{$matrix}). For details, see the "Getting Started" vignette.
}
\examples{
data(polymod)
pop <- wpp_age("United Kingdom", 2005)
polymod |>
  (\(s) s[country == "United Kingdom"])() |>
  assign_age_groups(age_limits = c(0, 5, 15)) |>
  compute_matrix() |>
  split_matrix(survey_pop = pop)

}
