% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.r
\name{survey_countries}
\alias{survey_countries}
\title{List all countries contained in a survey}
\usage{
survey_countries(survey, country.column = "country")
}
\arguments{
\item{survey}{survey to list countries from ("POLYMOD")}

\item{country.column}{column in the survey indicating the country}
}
\value{
list of countries
}
\description{
List all countries contained in a survey
}

