% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.r
\name{survey_countries}
\alias{survey_countries}
\title{List all countries contained in a survey}
\usage{
survey_countries(country.column = "country", ...)
}
\arguments{
\item{country.column}{column in the survey indicating the country}

\item{...}{arguments for \link{\code{get_survey}}, especially \code{survey}}

\item{survey}{survey to list countries from ("POLYMOD")}
}
\value{
list of countries
}
\description{
List all countries contained in a survey
}
