% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.r
\name{survey_countries}
\alias{survey_countries}
\title{List all countries contained in a survey}
\usage{
survey_countries(survey, country.column = "country", ...)
}
\arguments{
\item{survey}{a DOI (see \code{\link[=list_surveys]{list_surveys()}}), or a character vector of file names, or a \code{\link[=survey]{survey()}} object (in which case only cleaning is done).}

\item{country.column}{column in the survey indicating the country}

\item{...}{further arguments for \code{\link[=get_survey]{get_survey()}}}
}
\value{
list of countries
}
\description{
List all countries contained in a survey
}
\examples{
data(polymod)
survey_countries(polymod)
}
