% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_countries.r, R/wpp_age.r
\name{wpp_age}
\alias{wpp_age}
\alias{wpp_age}
\title{Find countries}
\usage{
wpp_age(countries, years)

wpp_age(countries, years)
}
\arguments{
\item{countries}{countries, will return all if not given}

\item{years}{years, will return all if not given}

\item{countries}{countries, will return all if not given}

\item{years}{years, will return all if not given}
}
\value{
data frame of age-specific population data

data frame of age-specific population data
}
\description{
This uses data from the \code{wpp2015} package but combines male and female,
and converts age groups to lower age limits

This uses data from the \code{wpp2015} package but combines male and female,
and converts age groups to lower age limits
}
