% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wpp_age.r
\name{wpp_age}
\alias{wpp_age}
\title{Get age-specific population data according to the World Population Prospects 2017 edition}
\usage{
wpp_age(countries, years)
}
\arguments{
\item{countries}{countries, will return all if not given}

\item{years}{years, will return all if not given}
}
\value{
data frame of age-specific population data
}
\description{
This uses data from the \code{wpp2017} package but combines male and female,
and converts age groups to lower age limits. If the requested year is not present 
in the historical data, wpp projections are used.
}
\examples{
wpp_age("Italy", c(1990, 2000))
}
