% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lists.r
\name{wpp_countries}
\alias{wpp_countries}
\title{List all countries and regions for which socialmixr has population data}
\usage{
wpp_countries()
}
\value{
list of countries
}
\description{
Uses the World Population Prospects data from the \code{wpp2015} package
}
\examples{
\dontrun{wpp_countries()}
}
